+++ .stylelintrc.json
@@ -34,6 +34,7 @@
     "**/reset*.css",
     "**/bootstrap*.css",
     "**/*.js",
-    "**/*.jsx"
+    "**/*.jsx",
+    "coverage/**"
   ]
 }
